+++ src/app/state/find/index.js
@@ -1,0 +1,9 @@
+export const initial = {
+	open: false,
+	query: '',
+	caseSensitive: false
+};
+
+export default {
+	initial
+};